Improve running integration tests (#61)
* Upgrade to latest Exasol version

* Adapt integration test for macOS

* Allow using a running database

* Add unit tests for JSON marshalling

* Cleanup schemas after tests

+++ integration_test.go
@@ -7,6 +7,7 @@ import (
 	"log"
 	"os"
 	"regexp"
+	"strconv"
 	"strings"
 	"testing"
 	"time"
@@ -35,9 +36,29 @@ func TestIntegrationSuite(t *testing.T) {
 
 func (suite *IntegrationTestSuite) SetupSuite() {
 	suite.ctx = getContext()
-	suite.exasolContainer = runExasolContainer(suite.ctx)
-	suite.port = getExasolPort(suite.exasolContainer, suite.ctx)
-	suite.host = getExasolHost(suite.exasolContainer, suite.ctx)
+	exasolHostEnv := os.Getenv("EXASOL_HOST")
+	if exasolHostEnv != "" {
+		suite.exasolContainer = nil
+		suite.host = exasolHostEnv
+		suite.port = getExasolPortFromEnv()
+	} else {
+		suite.exasolContainer = runExasolContainer(suite.ctx)
+		suite.port = getExasolPort(suite.exasolContainer, suite.ctx)
+		suite.host = getExasolHost(suite.exasolContainer, suite.ctx)
+	}
+}
+
+func getExasolPortFromEnv() int {
+	envValue := os.Getenv("EXASOL_PORT")
+	if envValue == "" {
+		envValue = "8563"
+	}
+	if intValue, err := strconv.Atoi(envValue); err == nil {
+		return intValue
+	} else {
+		log.Fatalf("Error parsing port %q from environment variable EXASOL_PORT", envValue)
+		return -1
+	}
 }
 
 func getExasolHost(exasolContainer testcontainers.Container, ctx context.Context) string {
@@ -70,7 +91,7 @@ func (suite *IntegrationTestSuite) TestConnection() {
 	if suite.host == "localhost" {
 		errorMsgCertWrongHost = "x509: certificate is not valid for any names, but wanted to match localhost"
 	} else {
-		errorMsgCertWrongHost = fmt.Sprintf("x509: cannot validate certificate for %s because it doesn't contain any IP SANs", suite.host)
+		errorMsgCertWrongHost = "x509: “*.exacluster.local” certificate is not standards compliant"
 	}
 	noError := ""
 
@@ -132,6 +153,7 @@ func (suite *IntegrationTestSuite) TestExecAndQuery() {
 	database := suite.openConnection(suite.createDefaultConfig())
 	schemaName := "TEST_SCHEMA_1"
 	_, _ = database.Exec("CREATE SCHEMA " + schemaName)
+	defer suite.dropSchema(database, schemaName)
 	_, _ = database.Exec("CREATE TABLE " + schemaName + ".TEST_TABLE(x INT)")
 	_, _ = database.Exec("INSERT INTO " + schemaName + ".TEST_TABLE VALUES (15)")
 	rows, _ := database.Query("SELECT x FROM " + schemaName + ".TEST_TABLE")
@@ -142,6 +164,7 @@ func (suite *IntegrationTestSuite) TestFetch() {
 	database := suite.openConnection(suite.createDefaultConfig().FetchSize(200))
 	schemaName := "TEST_SCHEMA_FETCH"
 	_, _ = database.Exec("CREATE SCHEMA " + schemaName)
+	defer suite.dropSchema(database, schemaName)
 	_, _ = database.Exec("CREATE TABLE " + schemaName + ".TEST_TABLE(x INT)")
 	data := make([]string, 0)
 	for i := 0; i < 10000; i++ {
@@ -173,6 +196,7 @@ func (suite *IntegrationTestSuite) TestQueryWithError() {
 	database := suite.openConnection(suite.createDefaultConfig())
 	schemaName := "TEST_SCHEMA_2"
 	_, _ = database.Exec("CREATE SCHEMA " + schemaName)
+	defer suite.dropSchema(database, schemaName)
 	_, err := database.Query("SELECT x FROM " + schemaName + ".TEST_TABLE")
 	suite.Error(err)
 	suite.True(strings.Contains(err.Error(), "object TEST_SCHEMA_2.TEST_TABLE not found"))
@@ -182,6 +206,7 @@ func (suite *IntegrationTestSuite) TestPreparedStatement() {
 	database := suite.openConnection(suite.createDefaultConfig())
 	schemaName := "TEST_SCHEMA_3"
 	_, _ = database.Exec("CREATE SCHEMA " + schemaName)
+	defer suite.dropSchema(database, schemaName)
 	_, _ = database.Exec("CREATE TABLE " + schemaName + ".TEST_TABLE(x INT)")
 	preparedStatement, _ := database.Prepare("INSERT INTO " + schemaName + ".TEST_TABLE VALUES (?)")
 	_, _ = preparedStatement.Exec(15)
@@ -195,6 +220,7 @@ func (suite *IntegrationTestSuite) TestBeginAndCommit() {
 	schemaName := "TEST_SCHEMA_4"
 	transaction, _ := database.Begin()
 	_, _ = transaction.Exec("CREATE SCHEMA " + schemaName)
+	defer suite.dropSchema(database, schemaName)
 	_, _ = transaction.Exec("CREATE TABLE " + schemaName + ".TEST_TABLE(x INT)")
 	_, _ = transaction.Exec("INSERT INTO " + schemaName + ".TEST_TABLE VALUES (15)")
 	_ = transaction.Commit()
@@ -207,6 +233,7 @@ func (suite *IntegrationTestSuite) TestBeginAndRollback() {
 	schemaName := "TEST_SCHEMA_5"
 	transaction, _ := database.Begin()
 	_, _ = transaction.Exec("CREATE SCHEMA " + schemaName)
+	defer suite.dropSchema(database, schemaName)
 	_, _ = transaction.Exec("CREATE TABLE " + schemaName + ".TEST_TABLE(x INT)")
 	_, _ = transaction.Exec("INSERT INTO " + schemaName + ".TEST_TABLE VALUES (15)")
 	_ = transaction.Rollback()
@@ -227,6 +254,7 @@ func (suite *IntegrationTestSuite) TestExecuteAndQueryWithContext() {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
 	schemaName := "TEST_SCHEMA_6"
 	_, _ = database.ExecContext(ctx, "CREATE SCHEMA "+schemaName)
+	defer suite.dropSchema(database, schemaName)
 	_, _ = database.ExecContext(ctx, "CREATE TABLE "+schemaName+".TEST_TABLE(x INT)")
 	_, _ = database.ExecContext(ctx, "INSERT INTO "+schemaName+".TEST_TABLE VALUES (15)")
 	rows, _ := database.QueryContext(ctx, "SELECT x FROM "+schemaName+".TEST_TABLE")
@@ -240,6 +268,7 @@ func (suite *IntegrationTestSuite) TestBeginWithCancelledContext() {
 	schemaName := "TEST_SCHEMA_7"
 	transaction, _ := database.BeginTx(ctx, nil)
 	_, _ = transaction.ExecContext(ctx, "CREATE SCHEMA "+schemaName)
+	defer suite.dropSchema(database, schemaName)
 	cancel()
 	_, err := transaction.ExecContext(ctx, "CREATE TABLE "+schemaName+".TEST_TABLE(x INT)")
 	suite.EqualError(err, "context canceled")
@@ -251,6 +280,7 @@ func (suite *IntegrationTestSuite) TestSimpleImportStatement() {
 	schemaName := "TEST_SCHEMA_8"
 	tableName := "TEST_TABLE"
 	_, _ = database.ExecContext(ctx, "CREATE SCHEMA "+schemaName)
+	defer suite.dropSchema(database, schemaName)
 	_, _ = database.ExecContext(ctx, fmt.Sprintf("CREATE TABLE %s.%s (a int , b VARCHAR(20))", schemaName, tableName))
 
 	result, err := database.ExecContext(ctx, fmt.Sprintf(`IMPORT INTO %s.%s FROM LOCAL CSV FILE './testData/data.csv' COLUMN SEPARATOR = ';' ENCODING = 'UTF-8' ROW SEPARATOR = 'LF'`, schemaName, tableName))
@@ -275,6 +305,7 @@ func (suite *IntegrationTestSuite) TestMultiImportStatement() {
 	schemaName := "TEST_SCHEMA_9"
 	tableName := "TEST_TABLE"
 	_, _ = database.ExecContext(ctx, "CREATE SCHEMA "+schemaName)
+	defer suite.dropSchema(database, schemaName)
 	_, _ = database.ExecContext(ctx, fmt.Sprintf("CREATE TABLE %s.%s (a int , b VARCHAR(20))", schemaName, tableName))
 
 	result, err := database.ExecContext(ctx, fmt.Sprintf(`IMPORT INTO %s.%s FROM LOCAL CSV FILE './testData/data.csv' FILE './testData/data_part2.csv' COLUMN SEPARATOR = ';' ENCODING = 'UTF-8' ROW SEPARATOR = 'LF'`, schemaName, tableName))
@@ -321,6 +352,7 @@ func (suite *IntegrationTestSuite) TestImportStatementWithCRFile() {
 	schemaName := "TEST_SCHEMA_10"
 	tableName := "TEST_TABLE"
 	_, _ = database.ExecContext(ctx, "CREATE SCHEMA "+schemaName)
+	defer suite.dropSchema(database, schemaName)
 	_, _ = database.ExecContext(ctx, fmt.Sprintf("CREATE TABLE %s.%s (a int , b VARCHAR(20))", schemaName, tableName))
 
 	result, err := database.ExecContext(ctx, fmt.Sprintf(`IMPORT INTO %s.%s FROM LOCAL CSV FILE './testData/data_cr.csv' COLUMN SEPARATOR = ';' ENCODING = 'UTF-8' ROW SEPARATOR = 'CR'`, schemaName, tableName))
@@ -347,9 +379,17 @@ func (suite *IntegrationTestSuite) assertSingleValueResult(rows *sql.Rows, expec
 	suite.Equal(expected, testValue)
 }
 
+func (suite *IntegrationTestSuite) dropSchema(db *sql.DB, schemaName string) {
+	_, err := db.Exec("DROP SCHEMA IF EXISTS " + schemaName + " CASCADE")
+	suite.NoError(err, "Failed to drop schema "+schemaName)
+	fmt.Printf("Schema dropped: %q\n", schemaName)
+}
+
 func (suite *IntegrationTestSuite) TearDownSuite() {
-	err := suite.exasolContainer.Terminate(suite.ctx)
-	onError(err)
+	if suite.exasolContainer != nil {
+		err := suite.exasolContainer.Terminate(suite.ctx)
+		onError(err)
+	}
 }
 
 func (suite *IntegrationTestSuite) createDefaultConfig() *exasol.DSNConfigBuilder {
@@ -373,7 +413,7 @@ func runExasolContainer(ctx context.Context) testcontainers.Container {
 
 	dbVersion := os.Getenv("DB_VERSION")
 	if dbVersion == "" {
-		dbVersion = "7.1.8"
+		dbVersion = "7.1.9"
 	}
 
 	request := testcontainers.ContainerRequest{
